third_party/blink/web_tests/media/controls/quirks-mode-timeline-is-not-hidden.html

<html>
<!-- We don't declare a DOCTYPE in order to force quirks mode -->
<title>Test that on quirks mode the timeline is not hidden.</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../media-controls.js"></script>
<body>
<video controls width=400 src="../content/60_sec_video.webm"></video>
<script>
async_test(t => {
  const video = document.querySelector('video');

  video.onloadedmetadata = t.step_func_done(() => {
    const timelineContainer = timelineElement(video);
    const timeline = internals.shadowRoot(timelineContainer).firstChild;
    const timelineHeight = getComputedStyle(timeline).getPropertyValue("height");
    assert_not_equals(timelineHeight, "0px");
  });
});
</script>
</body>
</html>
